refactor: move index router to file
